+++ src/yooai/containers/Pane.ts
@@ -1,5 +1,5 @@
 import {Container} from "pixi.js";
-import {Component} from "../..";
+import {Component, invalidate} from "../..";
 
 export class Pane extends Component {
 
@@ -7,13 +7,11 @@ export class Pane extends Component {
     return this._skin;
   }
 
+  @invalidate("skin")
   public set skin(value: Container | undefined) {
-    if (this._skin === value) {
-      return;
-    }
     this._skin = value;
-    this.invalidate("skin");
   }
+
   protected _skin?: Container;
   protected _background?: Container;
 
@@ -36,13 +34,7 @@ export class Pane extends Component {
   }
 
   protected drawBackground() {
-    if (this._background !== undefined && this._background.parent === this) {
-      this.removeChild(this._background);
-    }
-    this._background = this._skin;
-    if (this._background) {
-      this.addChildAt(this._background, 0);
-    }
+    this._background = this.updateSkin(this._background, this._skin, 0);
   }
 
   protected drawLayout() {