+++ src/yooai/layout/IHasDimensions.ts
@@ -1,0 +1,6 @@
+import {IPoint} from "./IPoint";
+
+export interface IHasDimensions extends Partial<IPoint> {
+  width: number;
+  height: number;
+}